+++ src/py42/services/detectionlists/high_risk_employee.py
@@ -1,4 +1,6 @@
 from py42.exceptions import Py42BadRequestError
+from py42.exceptions import Py42NotFoundError
+from py42.exceptions import Py42UserNotOnListError
 from py42.services import BaseService
 from py42.services.detectionlists import _DetectionListFilters
 from py42.services.detectionlists import _PAGE_SIZE
@@ -89,7 +91,14 @@ class HighRiskEmployeeService(BaseService):
             u"userId": user_id,
         }
         uri = self._make_uri(u"/remove")
-        return self._connection.post(uri, json=data)
+        try:
+            return self._connection.post(uri, json=data)
+        except Py42NotFoundError as err:
+            message = (
+                "User with ID '{}' is not currently on the high-risk-employee "
+                "detection list.".format(user_id)
+            )
+            raise Py42UserNotOnListError(err, message=message)
 
     def get(self, user_id):
         """Gets user information.