+++ CHANGELOG.md
@@ -18,6 +18,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 - Added new exception `Py42CloudAliasLimitExceededError` to throw if `add_cloud_alias()` throws `400` and body contains
 reason `Cloud usernames must be less than or equal to`.
+- Added new exception `Py42UserNotOnListError` to throw error if `remove()` throws `404` when the user is not on a detection list.
 
 ## 1.9.0 - 2020-10-02